+++ crates/orchestrator/src/tests/jobs/proving_job/mod.rs
@@ -1,8 +1,14 @@
+use bytes::Bytes;
 use std::collections::HashMap;
+use std::fs::File;
+use std::io::Read;
+use std::path::Path;
 use std::sync::Arc;
 
 use crate::config::config;
+use crate::data_storage::MockDataStorage;
 use httpmock::prelude::*;
+use mockall::predicate::eq;
 use prover_client_interface::{MockProverClient, TaskStatus};
 use rstest::*;
 use starknet::providers::jsonrpc::HttpTransport;
@@ -65,9 +71,19 @@ async fn test_process_job() {
         Url::parse(format!("http://localhost:{}", server.port()).as_str()).expect("Failed to parse URL"),
     ));
 
+    let mut file =
+        File::open(Path::new(&format!("{}/src/tests/artifacts/fibonacci.zip", env!("CARGO_MANIFEST_DIR")))).unwrap();
+    let mut buffer = Vec::new();
+    file.read_to_end(&mut buffer).unwrap();
+
+    let mut storage = MockDataStorage::new();
+    let buffer_bytes = Bytes::from(buffer);
+    storage.expect_get_data().with(eq("0/pie.zip")).return_once(move |_| Ok(buffer_bytes));
+
     TestConfigBuilder::new()
         .mock_starknet_client(Arc::new(provider))
         .mock_prover_client(Box::new(prover_client))
+        .mock_storage_client(Box::new(storage))
         .build()
         .await;